+++ packages/fx-core/src/common/featureFlags.ts
@@ -26,3 +26,15 @@ export function initializePreviewFeatureFlags(): void {
 export function isCLIDotNetEnabled(): boolean {
   return isFeatureFlagEnabled(FeatureFlagName.CLIDotNet, false);
 }
+
+export function isV3Enabled(): boolean {
+  return process.env.TEAMSFX_V3 ? process.env.TEAMSFX_V3 === "true" : true;
+}
+
+export function isVideoFilterEnabled(): boolean {
+  return isFeatureFlagEnabled(FeatureFlagName.VideoFilter, false);
+}
+
+export function isImportSPFxEnabled(): boolean {
+  return isFeatureFlagEnabled(FeatureFlagName.ImportSPFx, false);
+}
