+++ packages/fx-core/src/common/samples.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { sendRequestWithTimeout } from "../component/generator/utils";
 import sampleConfigV3 from "./samples-config-v3.json";
-import { isVideoFilterEnabled } from "./tools";
+import { isVideoFilterEnabled } from "./featureFlags";
 
 class configInfo {
   static readonly owner = "OfficeDev";
