+++ packages/fx-core/src/common/tools.ts
@@ -160,18 +160,6 @@ export function getResourceGroupInPortal(
   }
 }
 
-export function isV3Enabled(): boolean {
-  return process.env.TEAMSFX_V3 ? process.env.TEAMSFX_V3 === "true" : true;
-}
-
-export function isVideoFilterEnabled(): boolean {
-  return isFeatureFlagEnabled(FeatureFlagName.VideoFilter, false);
-}
-
-export function isImportSPFxEnabled(): boolean {
-  return isFeatureFlagEnabled(FeatureFlagName.ImportSPFx, false);
-}
-
 export function compileHandlebarsTemplateString(templateString: string, context: any): string {
   const template = Handlebars.compile(templateString);
   return template(context);
