+++ packages/fx-core/src/core/FxCoreImplementV3.ts
@@ -72,8 +72,7 @@ import { pathUtils } from "../component/utils/pathUtils";
 import { FileNotFoundError, InvalidProjectError, UserCancelError } from "../error/common";
 import { NoNeedUpgradeError } from "../error/upgrade";
 import { YamlFieldMissingError } from "../error/yml";
-import { QuestionNames } from "../question";
-import { getQuestionsForCreateProject } from "../question/create";
+import { QuestionNames, getQuestionsForCreateProject } from "../question";
 import { checkPermission, grantPermission, listCollaborator } from "./collaborator";
 import { InvalidInputError, ObjectIsUndefinedError } from "./error";
 import { TOOLS } from "./globalVars";
