+++ packages/fx-core/src/question/questionNames.ts
@@ -27,5 +27,22 @@ export enum QuestionNames {
   RepalceTabUrl = "tdp-tab-url",
   ValidateMethod = "validate-method",
   AppPackagePath = "appPackagePath",
+
   Features = "features",
+  SourceEnvName = "sourceEnvName",
+  TargetEnvName = "targetEnvName",
+  TargetResourceGroupName = "targetResourceGroupName",
+  NewResourceGroupName = "newResourceGroupName",
+  NewResourceGroupLocation = "newResourceGroupLocation",
+  NewTargetEnvName = "newTargetEnvName",
+  ExistingTabEndpoint = "existing-tab-endpoint",
+  TeamsAppManifestFilePath = "manifest-path",
+  LocalTeamsAppManifestFilePath = "local-manifest-path",
+  AadAppManifestFilePath = "manifest-file-path",
+  TeamsAppPackageFilePath = "app-package-file-path",
+  ConfirmManifest = "confirmManifest",
+  ConfirmLocalManifest = "confirmLocalManifest",
+  OutputZipPathParamName = "output-zip-path",
+  OutputManifestParamName = "output-manifest-path",
+  M365Host = "m365-host",
 }
